@@ -1,40 +1,34 @@
 <?php
 
-/**
- * cookiebar extension for Contao Open Source CMS
+/*
+ * Cookiebar extension for Contao Open Source CMS
  *
- * Copyright (C) 2013 Codefog
+ * Copyright (C) 2011-2018 Codefog
  *
- * @package cookiebar
- * @author  Codefog <http://codefog.pl>
- * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
- * @author  Didier Federer <d.federer@designpilot.ch>
- * @license LGPL
+ * @author  Codefog <https://codefog.pl>
+ * @author  Kamil Kuzminski <https://github.com/qzminski>
+ * @license MIT
  */
 
-
 /**
- * Fields
+ * Fields.
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable']   = array('Cookiebar aktivieren', 'Die Cookie Infromation Bar auf der Webseite anzeigen.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_message']  = array('Cookies Information', 'Bitte eine Kurzbeschreibung der Cookies eingeben.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo']   = array('Details Seite', 'Hier können Sie die Seite wählen, welche mehr Informationen über die Cookies enthält.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = array('Cookiebar Position', 'Hier können Sie die Position der Cookiebar wählen.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_button']   = array('Button Label', 'Bitte geben Sie das Label des Buttons ein (z.B. <em>Akzeptieren</em>).');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_includeCss']   = array('Standard-CSS einbinden', 'Die Cookiebar-Standard-Styles einbinden.');
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable'] = ['Cookiebar aktivieren', 'Die Cookie Infromation Bar auf der Webseite anzeigen.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_message'] = ['Cookies Information', 'Bitte eine Kurzbeschreibung der Cookies eingeben.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo'] = ['Details Seite', 'Hier können Sie die Seite wählen, welche mehr Informationen über die Cookies enthält.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = ['Cookiebar Position', 'Hier können Sie die Position der Cookiebar wählen.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_button'] = ['Button Label', 'Bitte geben Sie das Label des Buttons ein (z.B. <em>Akzeptieren</em>).'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_includeCss'] = ['Standard-CSS einbinden', 'Die Cookiebar-Standard-Styles einbinden.'];
 
-
-
-/**
+/*
  * Legends
  */
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_legend'] = 'Cookie Information';
 
-
-/**
+/*
  * Reference
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top']    = 'oben';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top'] = 'oben';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['bottom'] = 'unten';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['before_wrapper'] = 'Vor dem #wrapper Tag';
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end']       = 'Vor dem schließendem body Tag';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end'] = 'Vor dem schließendem body Tag';
