@@ -1,39 +1,33 @@
 <?php
 
-/**
- * cookiebar extension for Contao Open Source CMS
+/*
+ * Cookiebar extension for Contao Open Source CMS
  *
- * Copyright (C) 2013 Codefog
+ * Copyright (C) 2011-2018 Codefog
  *
- * @package cookiebar
- * @author  Codefog <http://codefog.pl>
- * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
- * @author  ContaoCms.it <http://contaocms.it>
- * @author  Paolo Brunelli <paolob@contaocms.it>
- * @license LGPL
+ * @author  Codefog <https://codefog.pl>
+ * @author  Kamil Kuzminski <https://github.com/qzminski>
+ * @license MIT
  */
 
-
 /**
- * Fields
+ * Fields.
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable']   = array('Abilita cookie bar', 'Mostra la barra delle informazioni dei cookies sul sito.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_message']  = array('Informazioni Cookies ', 'Inserire una breve descrizione sull\'utilizzo dei cookies.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo']   = array('Pagina dettaglio', 'Selezionare la pagina in cui sono presenti le informazioni relative ai cookies.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = array('Posizione', 'Indicare in quale posizione inserire la barra informazioni.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_button']   = array('Etichetta pulsante', 'Indicare l\'etichetta del pulsante (es. <em>Accetta</em>).');
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable'] = ['Abilita cookie bar', 'Mostra la barra delle informazioni dei cookies sul sito.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_message'] = ['Informazioni Cookies ', 'Inserire una breve descrizione sull\'utilizzo dei cookies.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo'] = ['Pagina dettaglio', 'Selezionare la pagina in cui sono presenti le informazioni relative ai cookies.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = ['Posizione', 'Indicare in quale posizione inserire la barra informazioni.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_button'] = ['Etichetta pulsante', 'Indicare l\'etichetta del pulsante (es. <em>Accetta</em>).'];
 
-
-/**
+/*
  * Legends
  */
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_legend'] = 'Informazioni Cookie';
 
-
-/**
+/*
  * Reference
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top']    = 'In alto';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top'] = 'In alto';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['bottom'] = 'In basso';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['before_wrapper'] = 'Prima #wrapper tag';
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end']       = 'Prima body Tag';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end'] = 'Prima body Tag';
