@@ -1,37 +1,33 @@
 <?php
 
-/**
- * cookiebar extension for Contao Open Source CMS
+/*
+ * Cookiebar extension for Contao Open Source CMS
  *
- * Copyright (C) 2013 Codefog
+ * Copyright (C) 2011-2018 Codefog
  *
- * @package cookiebar
- * @author  Codefog <http://codefog.pl>
- * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
- * @license LGPL
+ * @author  Codefog <https://codefog.pl>
+ * @author  Kamil Kuzminski <https://github.com/qzminski>
+ * @license MIT
  */
 
-
 /**
- * Fields
+ * Fields.
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable']   = array('Włącz komunikat o ciasteczkach', 'Wyświetla informację o ciasteczkach na stronie.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_message']  = array('Informacja o ciasteczkach', 'Wprowadź informację o ciasteczkach.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo']   = array('Strona ze szczegółami', 'Tutaj możesz wybrać stronę, na której znajduje się więcej szczegółów na temat ciasteczek.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = array('Pozycja komunikatu', 'Tutaj możesz wybrać pozycję komunikatu.');
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_button']   = array('Napis na przycisku', 'Wprowadź napis na przycisku (np. <em>Akceptuję</em>).');
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable'] = ['Włącz komunikat o ciasteczkach', 'Wyświetla informację o ciasteczkach na stronie.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_message'] = ['Informacja o ciasteczkach', 'Wprowadź informację o ciasteczkach.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo'] = ['Strona ze szczegółami', 'Tutaj możesz wybrać stronę, na której znajduje się więcej szczegółów na temat ciasteczek.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = ['Pozycja komunikatu', 'Tutaj możesz wybrać pozycję komunikatu.'];
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_button'] = ['Napis na przycisku', 'Wprowadź napis na przycisku (np. <em>Akceptuję</em>).'];
 
-
-/**
+/*
  * Legends
  */
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_legend'] = 'Informacje o cookie';
 
-
-/**
+/*
  * Reference
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top']    = 'Na górze strony';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top'] = 'Na górze strony';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['bottom'] = 'Na dole strony';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['before_wrapper'] = 'Przed #wrapper tag';
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end']       = 'Przed body Tag';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end'] = 'Przed body Tag';
