@@ -1,30 +1,27 @@
 <?php
 
-/**
- * cookiebar extension for Contao Open Source CMS
+/*
+ * Cookiebar extension for Contao Open Source CMS
  *
- * Copyright (C) 2011-2017 Codefog
+ * Copyright (C) 2011-2018 Codefog
  *
- * @package cookiebar
- * @author  Codefog <http://codefog.pl>
- * @author  Kamil Kuzminski <kamil.kuzminski@codefog.pl>
- * @author  Jan Kout <jan.kout@me.com>
- * @license LGPL
+ * @author  Codefog <https://codefog.pl>
+ * @author  Kamil Kuzminski <https://github.com/qzminski>
+ * @license MIT
  */
 
-
 /**
- * Fields
+ * Fields.
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable']   = [
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_enable'] = [
     'Aktivovat upozornění pro cookie',
     'Zobrazit informační box na webové stránce.',
 ];
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_message']  = [
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_message'] = [
     'Text upozornění o používání cookies',
     'Zadejte prosím krátký popisek cookies.',
 ];
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo']   = [
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_jumpTo'] = [
     'Stránka s podrobnějšími informacemi',
     'Zde můžete vybrat stránku, na kterou bude uživatel přesměrován pro další informace.',
 ];
@@ -32,22 +29,20 @@ $GLOBALS['TL_LANG']['tl_page']['cookiebar_position'] = [
     'Pozice upozornění pro cookie',
     'Zde můžete vybrat umístění upozornění pro cookie.',
 ];
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_button']   = [
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_button'] = [
     'Popisek tlačítka',
     'Zadejte prosím popisek tlačítka (z.B. <em>souhlasím</em>).',
 ];
 
-
-/**
+/*
  * Legends
  */
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_legend'] = 'Cookie Information';
 
-
-/**
+/*
  * Reference
  */
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top']             = 'nahoře';
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['bottom']          = 'dole';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['top'] = 'nahoře';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_position']['bottom'] = 'dole';
 $GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['before_wrapper'] = 'před tagem #wrapper';
-$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end']       = 'před zakončovacím tagem body';
+$GLOBALS['TL_LANG']['tl_page']['cookiebar_placement']['body_end'] = 'před zakončovacím tagem body';
